Revert "-Added TextConstantSizeComponent which shrinks as we add more letters to the text and vice versa. It's used in UIInputBoxes and buttons."
This reverts commit 6cb518cf6f1ba2e2d9154cb22a1db2206c47964e.

+++ Source/include/UI/Font.h
@@ -6,9 +6,9 @@
 struct Character
 {
 	unsigned int ID;
-	glm::vec2 Size;
-	glm::vec2 Bearing;
-	float Advance;
+	glm::ivec2 Size;
+	glm::ivec2 Bearing;
+	unsigned int Advance;
 };
 
 class Font
